Bump notepad== to latest in dev git repo
