+++ src/notepadequalequal/common.py
@@ -24,19 +24,6 @@ log_file = os.path.join(logDir, "notepadee_log.txt")
 # Get current PID
 pid = os.getpid()
 
-# Redirect output to log file
-log_file = open(log_file, 'a', encoding='utf-8', buffering=1)
-sys.stdout = log_file
-sys.stderr = log_file
-original_print = builtins.print
-def flushed_print(*args, **kwargs):
-    if 'flush' not in kwargs:
-        kwargs['flush'] = True
-    if 'end' not in kwargs:
-        kwargs['end'] = '\n'
-    return original_print(*args, **kwargs)
-builtins.print = flushed_print
-
 # Special printlog statement to print stuff that doesn't belong in a console to the log file
 def printlog(message, *args, **kwargs):
     # with open(log_file, 'a', encoding='utf-8') as file:
